requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
influxdb==5.3.1
influxdb-client==1.19.0
msgpack==1.0.2
numpy==1.21.0
pandas==1.3.0
psutil==5.8.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Rx==3.2.0
six==1.16.0
sqlparse==0.4.1
timeloop==1.0.2
urllib3==1.26.6
